Support meld as well as vimdiff for merging
When using meld, the left side has the "other file" i.e. the file with
the .rpmnew or .rpmsave extension. The right side has the "final file"
i.e. the target file into which changes have to be merged.

If the rpm action was to save the existing conf file, then the left side
will be the original conf file (.rpmsave) and the right side will be the
new rpm-based configuration. The user should merge any changes they want
to keep from their old .rpmsave from left to right.

If the rpm action was to keep the existing conf file, then the left side
will be the new conf file (.rpmnew) and right side will be the current
local configuration. The user should merge any changes in the new rpm
file into their local config from left to right.

Ideally these operations would be 3-way diffs between the local file and
the new rpm config, using the original rpm config as the common merge base.
However, the original rpm config is not available to yum and therefore a
3-way diff is not possible. Perhaps a future enhancement would be to use a
.orig file, if present, as the merge base for meld.

When using vimdiff, the order has not changed. The first parameter to
vimdiff is the "final: file i.e. the target file, and the second parameter
is the "other" file i.e. the .rpmnew or .rpmsave file.