Fix multiple non-spec argument processing (BZ 855239)
I doubt this ever actually worked, yum's bestPackagesFromList() appears
to be intended for picking the best packages from a set of related
packages, and in this case they're not even remotely related. Calling
bestPackagesFromList() separately for each package fixes things for
both src.rpm's in repositories and local src.rpms.