Revert "fs-snapshot: "dmsetup -o" workaround.  BZ 954358"
This reverts commit 0c0ddb41f347e9cdf9d3ea124fd87fd28b093a16.

Karel Srot from QA convinced me that explicit column list is
better than assuming the first two columns are ones we need.
Missing support for the "-o" option was just a regression.
It's safe to assume it's there to stay.

Note that the "-c" option is still required, so we should keep
commit 21fc32737ff3838c077fa36fdaeefda0ac5cbac2.