fastestmirror: move the logic before MD retrieval. BZ 1428210
This plugin has had no effect on metadata downloads since commit a522869
in yum where we put a RepoStorage.retrieveAllMD() call before
postreposetup_hook used by this plugin.  Let's change to using
prereposetup_hook instead so that retrieveAllMD() can make use of the
sorted urls.

In practice, though, mirrors would still be benchmarked and sorted
because of a similar mechanism present in urlgrabber (the timedhosts
option, always enabled in yum) with one exception:  If the yum cache is
empty (no timedhosts file yet), the original order received from the
server would be retained (a property of stable sorting [1]) for the
repomd.xml download, which is still a bad situation that wouldn't happen
before a522869; this is fixed with this commit as it allows the plugin
to do its work beforehand.

Note that the stuff we do in yum/repos.py in between prereposetup_hook
and postreposetup_hook has nothing to do with our plugin anyway, so we
won't "miss" or break any configuration steps when we move our logic
from the latter to the former location.

[1] urlgrabber/mirror.py:275