transaction-actions: fix file glob matching. BZ 1127783
Since one file shouldn't be provided by multiple packages, we can just
stop at the first matching package when dealing with an action key
that's a filename.  This is already handled correctly.

However, that's not desirable when dealing with a glob filename, in
which case multiple packages may have a file that matches that glob and
we should parse the action for each of those packages (this is important
because the action command may contain a variable that expands to the
respective package name/version/...).  This commit makes that happen.
Note that glob pkg specs are handled this way already.