versionlock: add hint and "status" subcommand. BZ 1497351
It is nice to be reminded of any versionlocks in effect when running yum
transactions (esp. when doing a "yum update" and wondering why no
updates were found).  This commit adds the following message that is
printed in the excludes hook:

  Excluding X updates due to versionlock (use "yum versionlock status"
  to show them)

Since it would be too annoying to have all the excludes printed every
time, we only show their count (X) and provide the new "status"
subcommand that allows the user to list them if they are interested.

Note that even this short message might be annoying as it appears on
every yum run, so we're also adding an option to disable it.